move distinct tests into their own files to ease reading
