+++ tests/parallel/main.rs
@@ -1,3 +1,5 @@
 mod setup;
 
 mod cancellation;
+mod independent;
+mod race;
